+++ src/main/java/com/nhnacademy/bookstoreorderapi/common/exception/ExternalServiceException.java
@@ -1,4 +1,4 @@
-package com.nhnacademy.bookstoreorderapi.order.client;
+package com.nhnacademy.bookstoreorderapi.common.exception;
 
 public class ExternalServiceException extends RuntimeException {
     public ExternalServiceException(String errorMessage, Throwable t) {
