+++ src/main/java/com/nhnacademy/bookstoreorderapi/order/client/user/exception/NotAdminException.java
@@ -1,4 +1,4 @@
-package com.nhnacademy.bookstoreorderapi.order.client;
+package com.nhnacademy.bookstoreorderapi.order.client.user.exception;
 
 public class NotAdminException extends RuntimeException {
     public NotAdminException(String message) {
